requirements.txt
flask==2.3.3
torch==2.9.1
torchvision==0.20.1
pillow==10.0.0
numpy==1.24.3
matplotlib==3.7.2
scikit-learn==1.3.0
